move hint register to circuit-std-rs/utils
