+++ circuit-std-rs/tests/gnark/emulated/field_bls12381/e2.rs
@@ -1,8 +1,7 @@
-use circuit_std_rs::gnark::{
+use circuit_std_rs::{gnark::{
     element::new_internal_element,
     emulated::field_bls12381::e2::{Ext2, GE2},
-    hints::register_hint,
-};
+}, utils::register_hint};
 use expander_compiler::frontend::compile_generic;
 use expander_compiler::{
     compile::CompileOptions,
