+++ circuit-std-rs/tests/gnark/emulated/field_bls12381/e6.rs
@@ -1,11 +1,10 @@
-use circuit_std_rs::gnark::{
+use circuit_std_rs::{gnark::{
     element::new_internal_element,
     emulated::field_bls12381::{
         e2::GE2,
         e6::{Ext6, GE6},
     },
-    hints::register_hint,
-};
+}, utils::register_hint};
 use expander_compiler::{
     compile::CompileOptions,
     declare_circuit,
