+++ circuit-std-rs/tests/gnark/emulated/sw_bls12381/pairing.rs
@@ -1,11 +1,10 @@
-use circuit_std_rs::gnark::{
+use circuit_std_rs::{gnark::{
     element::Element,
     emulated::{
         field_bls12381::e2::GE2,
         sw_bls12381::{g1::*, g2::*, pairing::Pairing},
     },
-    hints::register_hint,
-};
+},utils::register_hint};
 use expander_compiler::{
     declare_circuit,
     frontend::{extra::debug_eval, GenericDefine, HintRegistry, M31Config, RootAPI, Variable, M31},
